moved to Factory Method Pattern
