+++ src/AbstractValidator.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace DocumentValidator;
+
+abstract class AbstractValidator
+{
+    protected $docNumber;
+
+    public function __construct($docNumber)
+    {
+        $this->docNumber = $docNumber;
+    }
+
+    abstract public function isValid(): bool;
+    abstract protected function isValidFormat($docNumber): bool;
+
+    /*
+     *   This function validates the format of a given string in order to
+     *   see if it fits a regexp pattern.
+     *
+     *   This function is intended to work with Spanish identification
+     *   numbers, so it always checks string length (should be 9) and
+     *   accepts the absence of leading zeros.
+     *
+     *   This function returns:
+     *       TRUE: If specified string respects the pattern
+     *       FALSE: Otherwise
+     *
+     *   Usage:
+     *       echo respectsDocPattern(
+     *           '33576428Q',
+     *           '/^[KLM0-9][0-9][0-9][0-9][0-9][0-9][0-9][0-9][A-Z]/' );
+     *   Returns:
+     *       TRUE
+     */
+    protected function respectsDocPattern($givenString, $pattern): bool
+    {
+        $isValid = FALSE;
+        $fixedString = strtoupper($givenString);
+        if (is_int(substr($fixedString, 0, 1))) {
+            $fixedString = substr("000000000" . $givenString, -9);
+        }
+        if (preg_match($pattern, $fixedString)) {
+            $isValid = TRUE;
+        }
+        return $isValid;
+    }
+}
